fix: warning: declaration of 'lock' shadows a previous local [-Wshadow]
I swore I submitted this as a PR, but for the life of me I cannot find it.  This removes a warning for a lock that is shadowed.  In theory I shouldn't be an issue, but I noticed a few errors in helgrind warning about mutex locks forming out of order.  99.9% sure its boost asio, but this caught my eye.